Deep Neural Networks/src/requirements.txt
matplotlib
numpy
pillow
# tensorflow